refactor: 추천 여행지 상세 페이지 3단 분리 (Page/Container/Component)
